+++ tarpc/src/lib.rs
@@ -270,6 +270,7 @@ pub use crate::transport::sealed::Transport;
 use anyhow::Context as _;
 use futures::task::*;
 use std::{error::Error, fmt::Display, io, time::SystemTime};
+use std::sync::Arc;
 
 /// A message from a client to a server.
 #[derive(Debug)]
@@ -346,6 +347,29 @@ impl ServerError {
     pub fn new(kind: io::ErrorKind, detail: String) -> ServerError {
         Self { kind, detail }
     }
+}
+
+/// Critical errors that result in a Channel disconnecting.
+#[derive(thiserror::Error, Debug, PartialEq, Eq)]
+pub enum ChannelError<E>
+    where
+        E: Error + Send + Sync + 'static,
+{
+    /// Could not read from the transport.
+    #[error("could not read from the transport")]
+    Read(#[source] Arc<E>),
+    /// Could not ready the transport for writes.
+    #[error("could not ready the transport for writes")]
+    Ready(#[source] E),
+    /// Could not write to the transport.
+    #[error("could not write to the transport")]
+    Write(#[source] E),
+    /// Could not flush the transport.
+    #[error("could not flush the transport")]
+    Flush(#[source] E),
+    /// Could not close the write end of the transport.
+    #[error("could not close the write end of the transport")]
+    Close(#[source] E),
 }
 
 impl<T> Request<T> {
